Board creation, Deck creation
Board gets information from the database, this consists of both player buildings information.

The deck does the same but you can press on your card to get a prompt.

Known issues:
Board dissapears when its not your turn. (might need to change how the logic works on Action.js or inside their class)

Cards do the same but it is mostly intended.
Cards do not execute a play command.